requirements/requirements-packaging.txt
tensorflow>=2.2.0,<2.3.0
dlib==19.17.0
imutils>=0.5,<1.0
opencv-python>=4.2.0,<4.3.0
matplotlib>=3.2.0,<3.3.0
pillow>=7.1.0,<7.2.0
